# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CinemaVault API",
    "version": "0.6.0",
    "description": "Secure film discovery REST API for Coventry University 6003CEM CW2. It includes JWT authentication, RBAC, Basic Auth evidence, public film browsing, tracking records, direct messages, OMDB import, and OpenAPI documentation.",
    "contact": {
      "name": "CinemaVault coursework maintainer",
      "email": "placeholder@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v1/films": {
      "get": {
        "tags": ["Films"],
        "summary": "Browse, search, filter, sort, and paginate live films",
        "parameters": [
          { "name": "title", "in": "query", "schema": { "type": "string" } },
          { "name": "genre", "in": "query", "schema": { "type": "string" } },
          { "name": "year", "in": "query", "schema": { "type": "integer" } },
          { "name": "rating", "in": "query", "schema": { "type": "number" } },
          { "name": "isLive", "in": "query", "schema": { "type": "boolean" } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "enum": ["title", "genre", "year", "rating", "createdAt", "updatedAt"] } },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "Film list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FilmListResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "post": {
        "tags": ["Films"],
        "summary": "Create a film manually",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FilmCreateRequest" },
              "example": {
                "title": "Inception",
                "genre": "Sci-Fi",
                "year": 2010,
                "rating": 8.8,
                "director": "Christopher Nolan",
                "isLive": true
              }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/FilmResponse" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/films/import-omdb": {
      "post": {
        "tags": ["Films"],
        "summary": "Import or update a film from OMDB",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OmdbImportRequest" },
              "examples": {
                "byImdbId": { "value": { "imdbId": "tt0133093" } },
                "byTitle": { "value": { "title": "The Matrix" } }
              }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/FilmResponse" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "502": { "$ref": "#/components/responses/BadGateway" }
        }
      }
    },
    "/api/v1/films/{id}": {
      "get": {
        "tags": ["Films"],
        "summary": "Get one film by id",
        "parameters": [{ "$ref": "#/components/parameters/FilmId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/FilmResponse" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "put": {
        "tags": ["Films"],
        "summary": "Update a film",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/FilmId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FilmUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/FilmResponse" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "delete": {
        "tags": ["Films"],
        "summary": "Soft-delete a film",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/FilmId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/FilmResponse" },
          "204": { "description": "Deleted with no response body" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "parameters": {
      "FilmId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Forbidden": {
        "description": "Authenticated user lacks the required role",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Conflict": {
        "description": "Duplicate resource conflict",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "InternalServerError": {
        "description": "Unexpected server error",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "BadGateway": {
        "description": "External OMDB API failed",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "FilmResponse": {
        "description": "Film response",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Film" } } }
      },
...
    },
    "schemas": {
...
      "Film": {
        "type": "object",
        "required": ["id", "title", "isLive", "createdAt", "updatedAt", "links"],
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "genre": { "type": "string", "nullable": true },
          "year": { "type": "integer", "nullable": true },
          "rating": { "type": "number", "nullable": true },
          "director": { "type": "string", "nullable": true },
          "cast": { "type": "string", "nullable": true },
          "plot": { "type": "string", "nullable": true },
          "posterUrl": { "type": "string", "nullable": true },
          "runtime": { "type": "integer", "nullable": true },
          "language": { "type": "string", "nullable": true },
          "country": { "type": "string", "nullable": true },
          "imdbId": { "type": "string", "nullable": true },
          "omdbMetadataJson": { "type": "string", "nullable": true },
          "isLive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "links": { "type": "object", "additionalProperties": { "type": "string" } }
        }
      },
      "FilmCreateRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string" },
          "genre": { "type": "string" },
          "year": { "type": "integer", "minimum": 1888, "maximum": 2100 },
          "rating": { "type": "number", "minimum": 0, "maximum": 10 },
          "director": { "type": "string" },
          "cast": { "type": "string" },
          "plot": { "type": "string" },
          "posterUrl": { "type": "string", "format": "uri" },
          "runtime": { "type": "integer" },
          "language": { "type": "string" },
          "country": { "type": "string" },
          "imdbId": { "type": "string" },
          "omdbMetadataJson": { "type": "string" },
          "isLive": { "type": "boolean" }
        }
      },
      "FilmUpdateRequest": {
        "allOf": [{ "$ref": "#/components/schemas/FilmCreateRequest" }],
        "description": "Any non-empty subset of FilmCreateRequest fields."
      },
      "FilmListResponse": {
        "type": "object",
        "required": ["data", "pagination", "links"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Film" } },
          "pagination": {
            "type": "object",
            "properties": {
              "page": { "type": "integer" },
              "limit": { "type": "integer" },
              "total": { "type": "integer" },
              "totalPages": { "type": "integer" }
            }
          },
          "links": { "type": "object", "additionalProperties": { "type": "string" } }
        }
      },
      "OmdbImportRequest": {
        "type": "object",
        "properties": {
          "imdbId": { "type": "string" },
          "title": { "type": "string" }
        },
        "description": "Supply either imdbId or title."
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" },
          "details": { "type": "object" }
        }
      }
    }
  }
}